Extract event queue logic to separate file
Note no refactoring was done (yet), all just pure moving code around

This might break reference comparisons to __NOOP objects, will fix
